Fix Breaking Monuments in Spectator Post-Death.
Checks to see if player is considered dead by TGM, and if the block being broken is within a monument region, if the player is within x blocks of their spawn.

Root cause analysis still ongoing and this commit will be reverted when the cause of the BlockBreakEvents being triggered is found.