Fix Breaking Monuments in Spectator Post-Death
It was possible to be mining a monument, die, and then continue mining when put into spectator state (which is still in survival mode, for movement restriction, visibility restriction, and legacy code support).

It is still possible to trigger this big if the BlockBreakEvent is triggered as the player is respawning and no longer considered dead as soon as they respawn. This could be temporarily patched by adding new APIs to region to get the distance from edge of region to player location when BlockBreakEvent is triggered and if > someConstant, canceling it. Root cause analysis would be preferred and this condition is rare enough that it would be wise to leave it unaddressed from now and work on other issues.